backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=6.0.2",
    "django-rest-framework>=0.1.0",
    "pillow>=12.1.0",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
